refactor: extract test class
