+++ src/test/java/org/cicirello/search/problems/LCSTests.java
@@ -122,172 +122,6 @@ public class LCSTests {
   }
 
   @Test
-  public void testIsomorphicCase() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 0.5, true);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(minCost, problem.minCost());
-    assertEquals(maxValue, problem.maxValue());
-  }
-
-  @Test
-  public void testIsomorphicCaseSeed() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 0.5, true, 42);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(minCost, problem.minCost());
-    assertEquals(maxValue, problem.maxValue());
-  }
-
-  @Test
-  public void testIsomorphicCaseFullyConnected() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 1.0, true);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n * (n - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n * (n - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testIsomorphicCaseFullyConnectedSeed() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 1.0, true, 42);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(n * (n - 1) / 2, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(n * (n - 1) / 2, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testIsomorphicCaseNoEdges() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 0.0, true);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
-  public void testIsomorphicCaseNoEdgesSeed() {
-    int n = 7;
-    LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 0.0, true, 42);
-    assertEquals(n, problem.size());
-    Permutation p = new Permutation(n, 0);
-    int minCost = Integer.MAX_VALUE;
-    int maxCost = Integer.MIN_VALUE;
-    int minValue = Integer.MAX_VALUE;
-    int maxValue = Integer.MIN_VALUE;
-    Iterator<Permutation> iter = p.iterator();
-    while (iter.hasNext()) {
-      Permutation perm = iter.next();
-      int cost = problem.cost(perm);
-      int value = problem.value(perm);
-      if (cost < minCost) minCost = cost;
-      if (cost > maxCost) maxCost = cost;
-      if (value < minValue) minValue = value;
-      if (value > maxValue) maxValue = value;
-      assertEquals(problem.maxValue() - value, cost);
-    }
-    assertEquals(0, problem.minCost());
-    assertEquals(0, problem.maxValue());
-    assertEquals(0, minCost);
-    assertEquals(0, maxValue);
-    assertEquals(minCost, maxCost);
-    assertEquals(minValue, maxValue);
-  }
-
-  @Test
   public void testSameSizeRandom() {
     int n = 7;
     LargestCommonSubgraph problem = new LargestCommonSubgraph(n, 0.5, false);
